Move function into only file where it is referenced
